initialize: add project files with correct file architecture
This is the first commit for this project. Setup for main dependencies and files is complete. More packages will be installed in future commits as each file becomes populated.

Each file contains comments at the TOP of the file which contains the phrase "TODO:". This is how this project's tasks will be achieved, in addition to these commits.